pyproject.toml
[project]
name = "mcp-ai-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "google-adk>=1.1.1",
    "opik>= 1.7.28",
    "litellm>=1.72.0"
    ]
